+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- add option `DisableHTMLEscape` to parse log
+
 ## v2.0.3 - 02/10/2020
 
 ### Fixed